bug fix: exception handling

+++ backend/api.py
@@ -171,6 +171,12 @@ async def delete_table(project_id: str, token: str = Depends(get_current_token))
     """
     try:
         user_id = await get_user_id_from_token(token)
+        if user_id is None:
+            raise HTTPException(
+                status_code=HTTP_401_UNAUTHORIZED,
+                detail="Invalid or missing authorization token",
+                headers={"WWW-Authenticate": "Bearer"},
+            )
         status = await delete_project(project_id, user_id)
         return {"status": status}
     except Exception as error:
@@ -194,7 +200,6 @@ async def get_table(token: str = Depends(get_current_token)):
         A list of table names.
         :param token:
     """
-    status = 500
     try:
         user_id = await get_user_id_from_token(token)
         if user_id is None:
@@ -206,11 +211,10 @@ async def get_table(token: str = Depends(get_current_token)):
         result, status = await get_table_names(user_id)
         return {"result": result, "status": status}
     except Exception as error:
-        # raise HTTPException(
-        #     status_code=HTTP_500_INTERNAL_SERVER_ERROR, detail="Internal Server Error"
-        # ) from error
-        print(error)
-        return {"status": status}
+        raise HTTPException(
+            status_code=HTTP_500_INTERNAL_SERVER_ERROR,
+            detail="Internal Server Error"
+        ) from error
 
 
 @router.get("/api/getLineGraph/{api_token}/{graph}")
